requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
httpx>=0.27,<0.28
